+++ src/Minifier/MinifierInterface.php
@@ -20,9 +20,10 @@ interface MinifierInterface
 {
     public const TYPE_CSS = 'css';
     public const TYPE_JS = 'js';
+    public const TYPE_HTML = 'html';
 
     /**
-     * @param self::TYPE_CSS|self::TYPE_JS $type
+     * @param self::TYPE_CSS|self::TYPE_JS|self::TYPE_HTML $type
      */
-    public function minify(string $input, string $type): string;
+    public function minify(string $input, string $type/* , ?\Sensiolabs\MinifyBundle\Minifier\Options\OptionsInterface $options = null */): string;
 }