+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 1.3.0
+
+- Add HTML support (`MinifierInterface::TYPE_HTML`)
+- Add options support to `MinifierInterface::minify()` via the `OptionsInterface` DTO (BC-compatible, third argument)
+
+## 1.2.0
+
+- Add Symfony 8 support
+
 ## 1.1.0
 
 - Support for PHP 8.1.0 or higher